bugfix: Additional testing & bug fixes.

+++ src/domain/amount.rs
@@ -1,3 +1,3 @@
 // Amount in the form of integer to avoid float number
 // ex: 31234 -> $3.1234
-pub type Amount = i32;
+pub type Amount = i64;
